Working on conditional switching of cards
In this configuration, a new problem arises:
The program will not pick upo the card, even though the hand will include a 5-card combination if it did. It will not do it, because by picking it up, the chance for a good next card drops, bacause we have less combination end points. Need to chance the decision system, so it takes into consideration that longer combinations are more valuable.